Consolidate event listening into tokens service
This keeps the requests and the async websocket responses closer
together in the code.
