+++ src/tokens/tokens.interfaces.ts
@@ -1,21 +1,32 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsDefined, IsInt, IsNotEmpty, Min, NotContains } from 'class-validator';
 
-export class EthConnectAsyncResponse {
-  @ApiProperty()
+// Ethconnect interfaces
+export interface EthConnectAsyncResponse {
   sent: boolean;
-
-  @ApiProperty()
-  id: string;
-}
-
-export class AsyncResponse {
-  @ApiProperty()
   id: string;
 }
 
 export interface EthConnectReturn {
   output: string;
+}
+export interface UriEventData {
+  id: string;
+  value: string;
+}
+
+export interface TransferSingleEventData {
+  from: string;
+  to: string;
+  operator: string;
+  id: string;
+  value: number;
+}
+
+// REST API requests and responses
+export class AsyncResponse {
+  @ApiProperty()
+  id: string;
 }
 
 export enum TokenType {
@@ -99,4 +110,65 @@ export class TokenTransfer {
   @IsInt()
   @Min(1)
   amount: number;
+}
+
+// Websocket notifications
+
+export class TokenPoolEvent {
+  @ApiProperty()
+  pool_id: string;
+
+  @ApiProperty()
+  type: TokenType;
+
+  @ApiProperty()
+  namespace: string;
+
+  @ApiProperty()
+  name: string;
+
+  @ApiProperty()
+  client_id: string;
+}
+
+export class TokenMintEvent {
+  @ApiProperty()
+  pool_id: string;
+
+  @ApiProperty()
+  token_id: string;
+
+  @ApiProperty()
+  to: string;
+
+  @ApiProperty()
+  amount: number;
+}
+
+export class TokenTransferEvent {
+  @ApiProperty()
+  pool_id: string;
+
+  @ApiProperty()
+  token_id: string;
+
+  @ApiProperty()
+  from: string;
+
+  @ApiProperty()
+  to: string;
+
+  @ApiProperty()
+  amount: number;
+}
+
+export class ReceiptEvent {
+  @ApiProperty()
+  id: string;
+
+  @ApiProperty()
+  success: boolean;
+
+  @ApiProperty()
+  message?: string;
 }
