+++ src/tokens/tokens.module.ts
@@ -1,4 +1,5 @@
 import { HttpModule, Module } from '@nestjs/common';
+import { EventStreamProxyModule } from '../eventstream-proxy/eventstream-proxy.module';
 import { TokensController } from './tokens.controller';
 import { TokensService } from './tokens.service';
 @Module({
@@ -6,6 +7,7 @@ import { TokensService } from './tokens.service';
     HttpModule.register({
       timeout: 30000,
     }),
+    EventStreamProxyModule,
   ],
   controllers: [TokensController],
   providers: [TokensService],
